# Parallelize computing stats