--- a/pkg/segment/query/processor/statscommand_test.go
+++ b/pkg/segment/query/processor/statscommand_test.go
@@ -614,3 +614,284 @@ func Test_ProcessGroupByRequestFullBuffer(t *testing.T) {
 	// the p.bucketKeyWorkingBuf should be at least the size of the estimated buffer
 	assert.GreaterOrEqual(t, len(processor.bucketKeyWorkingBuf), estimatedSizeOfBuffer, "bucketKeyWorkingBuf size is less than estimated buffer size")
 }
+
+// Test_MergeIQRStatsResults_DifferentBuckets tests merging IQRs where each has completely different buckets.
+// IQR1 has buckets for keys: "A~X", "B~Y"
+// IQR2 has buckets for keys: "C~Z", "D~W"
+// Expected result: merged IQR should contain all 4 buckets with their individual stats.
+func Test_MergeIQRStatsResults_DifferentBuckets(t *testing.T) {
+	config.InitializeTestingConfig(t.TempDir())
+	config.GetRunningConfig().UseNewPipelineConverted = true
+
+	// Create IQR1 with data that produces buckets "A~X" and "B~Y"
+	iqr1 := iqr.NewIQR(0)
+	iqr1Data := map[string][]sutils.CValueEnclosure{
+		"category": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "A"},
+			{Dtype: sutils.SS_DT_STRING, CVal: "B"},
+		},
+		"type": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "X"},
+			{Dtype: sutils.SS_DT_STRING, CVal: "Y"},
+		},
+		"value": {
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(10)},
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(20)},
+		},
+	}
+	err := iqr1.AppendKnownValues(iqr1Data)
+	assert.NoError(t, err)
+
+	processor1 := NewStatsProcessor(&structs.StatsExpr{
+		GroupByRequest: &structs.GroupByRequest{
+			GroupByColumns: []string{"category", "type"},
+			MeasureOperations: []*structs.MeasureAggregator{
+				{MeasureCol: "value", MeasureFunc: sutils.Count},
+				{MeasureCol: "value", MeasureFunc: sutils.Sum},
+			},
+		},
+	})
+	processor1.setAsIqrStatsResults = true
+	_, err = processor1.Process(iqr1)
+	assert.NoError(t, err)
+	resultIqr1, err := processor1.Process(nil)
+	assert.Equal(t, io.EOF, err)
+
+	// Create IQR2 with data that produces buckets "C~Z" and "D~W"
+	iqr2 := iqr.NewIQR(0)
+	iqr2Data := map[string][]sutils.CValueEnclosure{
+		"category": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "C"},
+			{Dtype: sutils.SS_DT_STRING, CVal: "D"},
+		},
+		"type": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "Z"},
+			{Dtype: sutils.SS_DT_STRING, CVal: "W"},
+		},
+		"value": {
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(30)},
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(40)},
+		},
+	}
+	err = iqr2.AppendKnownValues(iqr2Data)
+	assert.NoError(t, err)
+
+	processor2 := NewStatsProcessor(&structs.StatsExpr{
+		GroupByRequest: &structs.GroupByRequest{
+			GroupByColumns: []string{"category", "type"},
+			MeasureOperations: []*structs.MeasureAggregator{
+				{MeasureCol: "value", MeasureFunc: sutils.Count},
+				{MeasureCol: "value", MeasureFunc: sutils.Sum},
+			},
+		},
+	})
+	processor2.setAsIqrStatsResults = true
+	_, err = processor2.Process(iqr2)
+	assert.NoError(t, err)
+	resultIqr2, err := processor2.Process(nil)
+	assert.Equal(t, io.EOF, err)
+
+	// Merge the IQRs
+	mergedIqr := iqr.NewIQR(0)
+	iqrs := []*iqr.IQR{resultIqr1, resultIqr2}
+	statsExists, err := mergedIqr.MergeIQRStatsResults(iqrs)
+	assert.NoError(t, err)
+	assert.True(t, statsExists)
+
+	// Verify merged results contain all 4 buckets
+	knownValues, err := mergedIqr.ReadAllColumns()
+	assert.NoError(t, err)
+
+	// Should have 4 rows - one for each bucket: A~X, B~Y, C~Z, D~W
+	assert.Len(t, knownValues["category"], 4)
+	assert.Len(t, knownValues["count(value)"], 4)
+	assert.Len(t, knownValues["sum(value)"], 4)
+
+	// Verify each bucket has count=1 and correct sum
+	expectedCounts := []uint64{1, 1, 1, 1}
+	expectedSums := []int64{10, 20, 30, 40}
+
+	actualCounts := make([]uint64, len(knownValues["count(value)"]))
+	actualSums := make([]int64, len(knownValues["sum(value)"]))
+
+	for i := range knownValues["count(value)"] {
+		actualCounts[i] = knownValues["count(value)"][i].CVal.(uint64)
+		actualSums[i] = knownValues["sum(value)"][i].CVal.(int64)
+	}
+
+	assert.ElementsMatch(t, expectedCounts, actualCounts)
+	assert.ElementsMatch(t, expectedSums, actualSums)
+}
+
+// Test_MergeIQRStatsResults_OverlappingBuckets tests merging IQRs where some buckets overlap.
+// IQR1 has buckets: "A~X" (count=1, sum=10), "B~Y" (count=1, sum=20)
+// IQR2 has buckets: "A~X" (count=1, sum=15), "C~Z" (count=1, sum=30)
+// Expected result: "A~X" combined (count=2, sum=25), "B~Y" unchanged, "C~Z" unchanged
+func Test_MergeIQRStatsResults_OverlappingBuckets(t *testing.T) {
+	config.InitializeTestingConfig(t.TempDir())
+	config.GetRunningConfig().UseNewPipelineConverted = true
+
+	// Create IQR1 with "A~X"=10, "B~Y"=20
+	iqr1 := iqr.NewIQR(0)
+	iqr1Data := map[string][]sutils.CValueEnclosure{
+		"category": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "A"},
+			{Dtype: sutils.SS_DT_STRING, CVal: "B"},
+		},
+		"type": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "X"},
+			{Dtype: sutils.SS_DT_STRING, CVal: "Y"},
+		},
+		"value": {
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(10)},
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(20)},
+		},
+	}
+	err := iqr1.AppendKnownValues(iqr1Data)
+	assert.NoError(t, err)
+
+	processor1 := NewStatsProcessor(&structs.StatsExpr{
+		GroupByRequest: &structs.GroupByRequest{
+			GroupByColumns: []string{"category", "type"},
+			MeasureOperations: []*structs.MeasureAggregator{
+				{MeasureCol: "value", MeasureFunc: sutils.Count},
+				{MeasureCol: "value", MeasureFunc: sutils.Sum},
+			},
+		},
+	})
+	processor1.setAsIqrStatsResults = true
+	_, err = processor1.Process(iqr1)
+	assert.NoError(t, err)
+	resultIqr1, err := processor1.Process(nil)
+	assert.Equal(t, io.EOF, err)
+
+	// Create IQR2 with "A~X"=15, "C~Z"=30 (overlaps with IQR1 on "A~X")
+	iqr2 := iqr.NewIQR(0)
+	iqr2Data := map[string][]sutils.CValueEnclosure{
+		"category": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "A"}, // Same as IQR1's first record
+			{Dtype: sutils.SS_DT_STRING, CVal: "C"}, // New bucket
+		},
+		"type": {
+			{Dtype: sutils.SS_DT_STRING, CVal: "X"}, // Same as IQR1's first record
+			{Dtype: sutils.SS_DT_STRING, CVal: "Z"}, // New bucket
+		},
+		"value": {
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(15)}, // Different value for A~X
+			{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(30)}, // New bucket value
+		},
+	}
+	err = iqr2.AppendKnownValues(iqr2Data)
+	assert.NoError(t, err)
+
+	processor2 := NewStatsProcessor(&structs.StatsExpr{
+		GroupByRequest: &structs.GroupByRequest{
+			GroupByColumns: []string{"category", "type"},
+			MeasureOperations: []*structs.MeasureAggregator{
+				{MeasureCol: "value", MeasureFunc: sutils.Count},
+				{MeasureCol: "value", MeasureFunc: sutils.Sum},
+			},
+		},
+	})
+	processor2.setAsIqrStatsResults = true
+	_, err = processor2.Process(iqr2)
+	assert.NoError(t, err)
+	resultIqr2, err := processor2.Process(nil)
+	assert.Equal(t, io.EOF, err)
+
+	// Merge the IQRs
+	mergedIqr := iqr.NewIQR(0)
+	iqrs := []*iqr.IQR{resultIqr1, resultIqr2}
+	statsExists, err := mergedIqr.MergeIQRStatsResults(iqrs)
+	assert.NoError(t, err)
+	assert.True(t, statsExists)
+
+	// Verify merged results
+	knownValues, err := mergedIqr.ReadAllColumns()
+	assert.NoError(t, err)
+
+	// Should have 3 unique buckets: A~X (merged), B~Y, C~Z
+	assert.Len(t, knownValues["category"], 3)
+
+	// Find the merged A~X bucket (should have count=2, sum=25)
+	// Find the unchanged B~Y bucket (should have count=1, sum=20)
+	// Find the new C~Z bucket (should have count=1, sum=30)
+	foundBuckets := make(map[string]struct {
+		count uint64
+		sum   int64
+	})
+
+	for i := 0; i < len(knownValues["category"]); i++ {
+		category := knownValues["category"][i].CVal.(string)
+		typeVal := knownValues["type"][i].CVal.(string)
+		count := knownValues["count(value)"][i].CVal.(uint64)
+		sum := knownValues["sum(value)"][i].CVal.(int64)
+
+		key := category + "~" + typeVal
+		foundBuckets[key] = struct {
+			count uint64
+			sum   int64
+		}{count: count, sum: sum}
+	}
+
+	// Verify expected buckets and their values
+	assert.Equal(t, uint64(2), foundBuckets["A~X"].count, "A~X should have count=2 (merged)")
+	assert.Equal(t, int64(25), foundBuckets["A~X"].sum, "A~X should have sum=25 (10+15)")
+
+	assert.Equal(t, uint64(1), foundBuckets["B~Y"].count, "B~Y should have count=1")
+	assert.Equal(t, int64(20), foundBuckets["B~Y"].sum, "B~Y should have sum=20")
+
+	assert.Equal(t, uint64(1), foundBuckets["C~Z"].count, "C~Z should have count=1")
+	assert.Equal(t, int64(30), foundBuckets["C~Z"].sum, "C~Z should have sum=30")
+}
+
+// Test_MergeIQRStatsResults_EmptyIQR tests merging when one IQR has no stats results.
+// IQR1 has stats, IQR2 is empty.
+// Expected result: should return the stats from IQR1 unchanged.
+func Test_MergeIQRStatsResults_EmptyIQR(t *testing.T) {
+	config.InitializeTestingConfig(t.TempDir())
+	config.GetRunningConfig().UseNewPipelineConverted = true
+
+	// Create IQR1 with data
+	iqr1 := iqr.NewIQR(0)
+	iqr1Data := map[string][]sutils.CValueEnclosure{
+		"category": {{Dtype: sutils.SS_DT_STRING, CVal: "A"}},
+		"value":    {{Dtype: sutils.SS_DT_SIGNED_NUM, CVal: int64(42)}},
+	}
+	err := iqr1.AppendKnownValues(iqr1Data)
+	assert.NoError(t, err)
+
+	processor1 := NewStatsProcessor(&structs.StatsExpr{
+		GroupByRequest: &structs.GroupByRequest{
+			GroupByColumns: []string{"category"},
+			MeasureOperations: []*structs.MeasureAggregator{
+				{MeasureCol: "value", MeasureFunc: sutils.Count},
+				{MeasureCol: "value", MeasureFunc: sutils.Sum},
+			},
+		},
+	})
+	processor1.setAsIqrStatsResults = true
+	_, err = processor1.Process(iqr1)
+	assert.NoError(t, err)
+	resultIqr1, err := processor1.Process(nil)
+	assert.Equal(t, io.EOF, err)
+
+	// Create empty IQR2
+	emptyIqr := iqr.NewIQR(0)
+
+	// Merge IQRs
+	mergedIqr := iqr.NewIQR(0)
+	iqrs := []*iqr.IQR{resultIqr1, emptyIqr}
+	statsExists, err := mergedIqr.MergeIQRStatsResults(iqrs)
+	assert.NoError(t, err)
+	assert.True(t, statsExists)
+
+	// Verify result is same as IQR1
+	knownValues, err := mergedIqr.ReadAllColumns()
+	assert.NoError(t, err)
+
+	assert.Len(t, knownValues["category"], 1)
+	assert.Equal(t, "A", knownValues["category"][0].CVal.(string))
+	assert.Equal(t, uint64(1), knownValues["count(value)"][0].CVal.(uint64))
+	assert.Equal(t, int64(42), knownValues["sum(value)"][0].CVal.(int64))
+}