# Description
When evaluating a query where `stats` is the first Bottleneck command, we now parallelize the stats computation across all the CPUs. This improves query performance significantly when there's only one query being executed.

As the image shows, we used to not parallelize the Data Processor chain; a recent PR made it so we parallelize up to but not including `stats`; now we parallelize `stats` and add another Data Processor to merge the stats.
![image](https://github.com/user-attachments/assets/89d76f00-e0b7-4cb9-8d54-b6977b65fda9)

# Testing
Tested with 100 million records on a few queries:
```
* | eval height=ResolutionHeight / 10 | stats avg(height) as avg, count(height) as count, sum(height) as sum by ResolutionWidth | sort -ResolutionWidth
```
This query time decreased from 25.5 seconds to 12.5 seconds. The results were the same except for some floating point arithmetic errors

```
Referer != "" | rex field=Referer "^https?://(?:www\.)?(?<k>[^/]+)" | stats avg(eval(len(Referer))) as l, count as c, min(eval(Referer)) by k | where c > 100000 | sort 25 -l
```
This query decreased from 22.4 to 13.2 seconds. This results in >10k buckets so we truncate to 10k, which causes the final results to differ between runs; but for the rows that were in both runs, the results are consistent